+++ resources/views/accounting/transactions/create.blade.php
@@ -35,6 +35,14 @@
                 {{ Form::bsFile('receipt_picture', [ 'accept' => 'image/*' ], __('accounting.choose_picture_of_receipt')) }}
             </div>
         </div>
+        <div class="form-row">
+            <div class="col-sm-4">
+                {{ Form::bsText('wallet_owner', Auth::user()->name, [ ], __('accounting.wallet_owner')) }}
+            </div>
+            <div class="col-sm-8">
+                {{ Form::bsText('remarks', null, [  ], __('app.remarks')) }}
+            </div>
+        </div>
         <p>
             {{ Form::bsSubmitButton(__('app.add')) }}
             <button type="submit" name="submit" value="save_and_continue" class="btn btn-secondary">@icon(arrow-right) @lang('app.save_and_continue')</button>